server/requirements.txt
fastapi
uvicorn
aioredis
aiohttp
python-dotenv
redis
